@@ -1,38 +1,46 @@
 import { Component } from '@angular/core';
-import { FooterComponent } from './footer.component';
 import { ReactiveFormsModule, FormGroup, FormControl, Validators } from '@angular/forms';
 import { CommonModule, JsonPipe } from '@angular/common';
 
 @Component({
-    selector: 'app-advanced-form',
-    imports: [CommonModule, ReactiveFormsModule, JsonPipe],
-    template: `
+  selector: 'app-advanced-form',
+  standalone: true,
+  imports: [CommonModule, ReactiveFormsModule, JsonPipe],
+  template: `
     <form [formGroup]="form" (ngSubmit)="onSubmit()" style="margin:1rem 0;">
       <label>
         Name:
         <input formControlName="name" />
       </label>
-      <div *ngIf="form.controls.name.invalid && form.controls.name.touched" style="color:red;">Name is required.</div>
+      @if (form.controls.name.invalid && form.controls.name.touched) {
+        <div style="color:red;">Name is required.</div>
+      }
       <br />
       <label>
         Email:
         <input formControlName="email" />
       </label>
-      <div *ngIf="form.controls.email.invalid && form.controls.email.touched" style="color:red;">Valid email required.</div>
+      @if (form.controls.email.invalid && form.controls.email.touched) {
+        <div style="color:red;">Valid email required.</div>
+      }
       <br />
       <label>
         Age:
         <input type="number" formControlName="age" />
       </label>
-      <div *ngIf="form.controls.age.invalid && form.controls.age.touched" style="color:red;">Age must be 18 or older.</div>
+      @if (form.controls.age.invalid && form.controls.age.touched) {
+        <div style="color:red;">Age must be 18 or older.</div>
+      }
       <br />
       <button type="submit" [disabled]="form.invalid">Submit</button>
     </form>
-    <div *ngIf="submitted">
-      <strong>Form Submitted!</strong>
-      <pre>{{ form.value | json }}</pre>
-    </div>
-  `
+    @if (submitted) {
+      <div>
+        <strong>Form Submitted!</strong>
+        <pre>{{ form.value | json }}</pre>
+      </div>
+    }
+    `
 })
 export class AdvancedFormComponent {
   form = new FormGroup({
