@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
@@ -13,21 +13,21 @@ export interface Contact {
 export class ContactsService {
   private apiUrl = 'http://localhost:3200/api/contacts';
 
-  constructor(private http: HttpClient) {}
+  private http = inject(HttpClient);
 
   getContacts(): Observable<Contact[]> {
     return this.http.get<Contact[]>(this.apiUrl);
   }
 
-  addContact(name: string, email: string, token: string): Observable<any> {
-    return this.http.post(this.apiUrl, { name, email, token });
+  addContact(name: string, email: string, token: string): Observable<Contact> {
+    return this.http.post<Contact>(this.apiUrl, { name, email, token });
   }
 
-  updateContact(id: number, name: string, email: string, token: string): Observable<any> {
-    return this.http.put(`${this.apiUrl}/${id}`, { name, email, token });
+  updateContact(id: number, name: string, email: string, token: string): Observable<Contact> {
+    return this.http.put<Contact>(`${this.apiUrl}/${id}`, { name, email, token });
   }
 
-  deleteContact(id: number, token: string): Observable<any> {
-    return this.http.delete(`${this.apiUrl}/${id}`, { body: { token } });
+  deleteContact(id: number, token: string): Observable<{ success: boolean }> {
+    return this.http.delete<{ success: boolean }>(`${this.apiUrl}/${id}`, { body: { token } });
   }
 }
