@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
@@ -12,17 +12,17 @@ export interface User {
 export class LoginService {
   private apiUrl = 'http://localhost:3200/api';
 
-  constructor(private http: HttpClient) {}
+  private http = inject(HttpClient);
 
-  login(username: string, password: string): Observable<any> {
-    return this.http.post(`${this.apiUrl}/login`, { username, password });
+  login(username: string, password: string): Observable<{ user: { username: string; role: string } }> {
+    return this.http.post<{ user: { username: string; role: string } }>(`${this.apiUrl}/login`, { username, password });
   }
 
-  register(username: string, password: string, role: string = 'user'): Observable<any> {
-    return this.http.post(`${this.apiUrl}/register`, { username, password, role });
+  register(username: string, password: string, role = 'user'): Observable<{ success: boolean }> {
+    return this.http.post<{ success: boolean }>(`${this.apiUrl}/register`, { username, password, role });
   }
 
-  delete(username: string): Observable<any> {
-    return this.http.post(`${this.apiUrl}/delete`, { username });
+  delete(username: string): Observable<{ success: boolean }> {
+    return this.http.post<{ success: boolean }>(`${this.apiUrl}/delete`, { username });
   }
 }
