@@ -3,13 +3,14 @@ import { RouterOutlet } from '@angular/router';
 import { HeaderComponent } from './header.component';
 import { MenuComponent } from './menu.component';
 import { FooterComponent } from './footer.component';
-import { CommonModule } from '@angular/common';
+
 
 @Component({
-    selector: 'app-root',
-    imports: [RouterOutlet, CommonModule, HeaderComponent, MenuComponent, FooterComponent],
-    templateUrl: './app.component.html',
-    styleUrl: './app.component.scss'
+  selector: 'app-root',
+  standalone: true,
+  imports: [RouterOutlet, HeaderComponent, MenuComponent, FooterComponent],
+  templateUrl: './app.component.html',
+  styleUrl: './app.component.scss'
 })
 export class AppComponent {
   title = 'angular-demo';
