@@ -7,9 +7,8 @@ import { GlobalErrorHandler } from './global-error-handler';
 import { provideRouter } from '@angular/router';
 import { provideClientHydration } from '@angular/platform-browser';
 import { routes } from './app.routes';
-import {} from '@angular/common/http';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 // ...existing code...
-import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { ErrorInterceptor } from './error.interceptor';
 
 export const appConfig: ApplicationConfig = {
