@@ -30,12 +30,12 @@ export class DataService {
     return this.http.get<User[]>(`${this.apiUrl}/users`);
   }
 
-  deleteUser(username: string): Observable<any> {
-    return this.http.post(`${this.apiUrl}/delete`, { username });
+  deleteUser(username: string): Observable<{ success: boolean }> {
+    return this.http.post<{ success: boolean }>(`${this.apiUrl}/delete`, { username });
   }
 
-  registerUser(username: string, password: string): Observable<any> {
-    return this.http.post(`${this.apiUrl}/register`, { username, password });
+  registerUser(username: string, password: string): Observable<{ success: boolean }> {
+    return this.http.post<{ success: boolean }>(`${this.apiUrl}/register`, { username, password });
   }
 
   getProducts(): Observable<Product[]> {
@@ -46,15 +46,15 @@ export class DataService {
     return this.http.get<Contact[]>(`${this.apiUrl}/contacts`);
   }
 
-  addContact(name: string, email: string, token: string): Observable<any> {
-    return this.http.post(`${this.apiUrl}/contacts`, { name, email, token });
+  addContact(name: string, email: string, token: string): Observable<{ success: boolean }> {
+    return this.http.post<{ success: boolean }>(`${this.apiUrl}/contacts`, { name, email, token });
   }
 
-  updateContact(id: number, name: string, email: string, token: string): Observable<any> {
-    return this.http.put(`${this.apiUrl}/contacts/${id}`, { name, email, token });
+  updateContact(id: number, name: string, email: string, token: string): Observable<{ success: boolean }> {
+    return this.http.put<{ success: boolean }>(`${this.apiUrl}/contacts/${id}`, { name, email, token });
   }
 
-  deleteContact(id: number, token: string): Observable<any> {
-    return this.http.delete(`${this.apiUrl}/contacts/${id}`, { body: { token } });
+  deleteContact(id: number, token: string): Observable<{ success: boolean }> {
+    return this.http.delete<{ success: boolean }>(`${this.apiUrl}/contacts/${id}`, { body: { token } });
   }
 }
