@@ -1,5 +1,5 @@
 import { Component, OnInit, inject } from '@angular/core';
-import { CommonModule } from '@angular/common';
+
 import { DataService, Product } from '../data.service';
 import { HeaderComponent } from '../header.component';
 import { MenuComponent } from '../menu.component';
@@ -7,7 +7,7 @@ import { FooterComponent } from '../footer.component';
 
 @Component({
     selector: 'app-products',
-    imports: [CommonModule, HeaderComponent, FooterComponent, MenuComponent],
+    imports: [HeaderComponent, FooterComponent, MenuComponent],
     templateUrl: './products.component.html',
     styleUrl: './products.component.scss'
 })
